No funciona aun, vamos a ver que pasa

+++ src/main/java/arcn/roomfinder/cliente/domain/model/CuentaBancaria.java
@@ -35,7 +35,7 @@ public class CuentaBancaria {
 
     public CuentaBancaria(CuentaBancariaEntidad cuentaBancariaEntidad) {
         this.numeroCuenta = cuentaBancariaEntidad.getNumeroCuenta();
-        this.cantidadCredito = cuentaBancariaEntidad.getCantidadCredito();
+        this.cantidadCredito = BigDecimal.valueOf(cuentaBancariaEntidad.getCantidadCredito());
     }
 
     public BigDecimal agregarFondos(BigDecimal cantidad){
